+++ src/Intermediate/Rxmxnx.PInvoke.Common.Intermediate/Internal/SuppressMessageConstants.cs
@@ -24,5 +24,10 @@ internal static class SuppressMessageConstants
 	public const String CheckIdS2696 = "S2696";
 	public const String CheckIdS3261 = "S3261";
 	public const String CheckIdS3218 = "S3218";
+	public const String CheckIdS3260 = "S3260";
+	public const String CheckIdS3776 = "S3776";
+	public const String CheckIdS1199 = "S1199";
+	public const String CheckIdS907 = "S907";
+	public const String CheckIdS1764 = "S1764";
 }
 #endif
